requirements.txt
# Core Web Framework
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2

# FastAPI for new VN-QUANT PRO
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
pydantic>=2.5.0

# Telegram Bot
python-telegram-bot[job-queue]

# Data & Visualization
plotly
requests
pandas
numpy
scipy
pytz
mplfinance

# Technical Analysis
ta>=0.11.0

# Backtesting & Statistics
statsmodels>=0.14.0

# Machine Learning (optional - for LSTM)
# tensorflow>=2.15.0
scikit-learn>=1.3.0

# Forecasting (optional - for Prophet)
# prophet>=1.1.0

# News & Sentiment
feedparser
beautifulsoup4
vaderSentiment

# Async Support
nest_asyncio
aiohttp>=3.9.0

# Utilities
python-dotenv
typing_extensions
